Give serves some grace period
If the AddRange can complete within the grace period, rangelet can
update the rangeInfo and return NsReady back to the controller in a
single RPC. If it takes longer, it can fall back to the previous
behavior of returning NsReadying.

This breaks orchestrator tests, because they expect the old behavior!
Fix before pushing.