Replace RsSubsuming with RsSplitting, RsJoining
The workflow to perform splits and joins is the same in many ways (it's
changing the boundaries of some ranges into some other boundaries of new
ranges), but the specifics can be quite different, especially under
complex error conditions like succesors crashing at inconvenient times.
There are a few places in which we're effectively inferring whether the
operation is a split or a join, so we might as well codify it.

Also, some clients are going to want to know whether they are part of a
split or a join, so it might pay to be explicit.

This does unfortunately preclude some interesting operations like
resharding two adjacent ranges into two adjacent ranges with a different
split point; a simultaneous split+join. But supporting that would be
hard for many clients anyway. We can add it later as a different
explicit operation.