Add IsParent, IsChild
This might prove to be pointless if we end up driving state changes from
within the operations, but right now they're on the keyspace, so are
driven in isolation. Each placement needs to figure out its role in the
operation (parent or child), so we provide these helpers, for now.