Refactor transitions to use dest/src
...as opposed to parent/child. That results in a lot of special-casing.
This is much simpler; we're just flowing from the source to the
destination, and sometimes (during error states) those invert.

The one remaining exception is in Drop. We generally want to drop
placements on the source-side once their replacements are activated on
the dest side, *and* drop them on the dest side if they've failed to
activate. But we specifically don't want to drop them in the case that
the dest side is the *parent* side, temporarily because one of the other
ranges on the child side failed to activate. In that case, we know that
we will retry soon, so want to leave it alone to avoid dropping and
re-placing the range which didn't fail. This is very convoluted, and I'm
still trying to figure out how to fit it into the src/dest model without
this special-casing.