Add Placement.Tainted
This is going to replace Placement.IsReplacing. The current approach
couples the old and new placements together during "moves", to help the
orchestrator understand that even though there's nothing wrong with the
current placement, it should deactivate to make way for the new one.

This is a relic from way back when operations were much more stateful.
It's awkward for many reasons, but mainly (a) it necessitates totally
separate codepaths in keyspace/operations.go for moves versus splits and
joins, and (b) it leaves the keyspace in a not broken but weird state
when either the old placement vanishes (the new one still refers to it
by NodeID, in IsReplacing) or when the new placement fails and is
destroyed (the move just never completes and the rpc hangs, but no
second attempt is made).

This revised approach basically makes the move operation two steps:
create a new placement, and taint the old one. The rest is automatic,
with no special concept of "replacement"; just like a split/join. But
it's not done yet! This only adds Tainted. Still have to remove
IsReplacing.