Replace replacedBy with Tainted
This one involves a slight behavior change. During a "move", when the
src placement failed/refused to deactivate, we used to drop the dest
which had just been placed (before deactivation was attempted). I was
trying to clean up the situation before getting wedged.

But in hindsight, that was wrong. The most likely thing to happen next
is for the src node to crash (because it's clearly fucked in some way)
or for an operator to forcibly drop the placement from it. At that
point, it's ideal to have the dest range already waiting in PsInactive,
rather than to start placement. This became obvious while I was
re-implementing the old behavior to make the tests pass.