!! Use 307 (Temporary Redirect) instead of 303 (Found)
This may be controversial. If this is a GET, there is no difference.
But HTTP clients that send other verbs (PUT, DELETE), on 303 may
send GET to the forwarded address. 307 explicitly requires the client
to send the same HTTP verb.

I don't know what the practical distinction is between 307 (temporary)
and 308 (permanent); I chose temporary because (if some balancer reassigns
the range) the resource could in principle end up back on this node.